backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.1.1
werkzeug==2.3.7
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.1.4
python-dotenv==1.0.0
